requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
tiktoken==0.4.0
faiss-cpu==1.7.4
protobuf~=3.19.0
google-generativeai
sentence_transformers
Flask==2.0.1

